requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
cryptography==3.4.7
Django==3.2.5
django-crispy-forms==1.12.0
django-embed-video==1.4.0
idna==3.2
pycparser==2.20
python-http-client==3.3.2
pytz==2021.1
requests==2.26.0
sendgrid==6.7.1
sqlparse==0.4.1
starkbank-ecdsa==1.1.1
typing-extensions==3.10.0.0
urllib3==1.26.6
